allow a space between the double slash and gcassert:
Previously, only the comments of the form `//gcassert:` would match the
regex pattern. This commit eases up the matching to also work for
`// gcassert:` (note the space after the double slash). We already have
a few cases like that in CockroachDB repo (even introduced by the author
of the gcassert himself :D ), so it makes sense to relax the matching
a bit.