fix: multiple funcion callsites on one line
Previously, if there were multiple callsites on one line, at least one
of which was a callsite for a //gcassert:inline directive but did not
actually get inlined; and another callsites on the same line *was*
inlined, the other callsite would be erroneously marked as having
succeeded to be inlined.

This is now corrected.